requirements (9).txt
flask 
tensorflow==2.15.1
werkzeug==2.0.3
numpy 
Pillow 
requests 
gdown 
gunicorn 

